Oops

+++ src/Servers/Kestrel/tools/CodeGenerator/KnownHeaders.cs
@@ -574,7 +574,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
         {(loop.ClassName == "HttpResponseHeaders" ? $@"
         internal unsafe void CopyToFast(ref BufferWriter<PipeWriter> output)
         {{
-            var tempBits = _bits | (_contentLength.HasValue ? {"0x" + (1L << 63).ToString("x")}L : 0);
+            var tempBits = (ulong)_bits | (_contentLength.HasValue ? {"0x" + (1L << 63).ToString("x")}L : 0);
             var next = 0;
             var keyStart = 0;
             var keyLength = 0;
